backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart
pydantic>=2.0.0
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv